+++ frontend/src/views/Settings.jsx
@@ -8,6 +8,7 @@ import {
     DangerZoneCard,
     ProjectUsers, ProjectCategories
 } from '../components/settings/settingsComponents';
+import {Header} from "../lib/ui/Header";
 
 const ProjectSettings = () => {
     const {
@@ -35,7 +36,9 @@ const ProjectSettings = () => {
 
     return (
         <div className="container mx-auto px-4 py-6">
-            <ProjectHeader
+            <Header
+                title="Project"
+                marked={"Settings"}
                 selectedProject={selectedProject}
                 loading={loading}
             />
