+++ frontend/src/views/Users.jsx
@@ -56,7 +56,6 @@ export const Users = () => {
 
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
                 {loading ? (
-                    // Render skeleton cards while loading
                     [...Array(6)].map((_, index) => (
                         <SkeletonCard key={index} header={true} lines={2} />
                     ))
